+++ tests/snapshot1.sh
@@ -21,7 +21,10 @@ test_pot_snapshot_001()
 	assertEquals "Help calls" "1" "$HELP_CALLS"
 	assertEquals "Error calls" "1" "$ERROR_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
+	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 
 	setUp
 	pot-snapshot -b bb
@@ -29,7 +32,10 @@ test_pot_snapshot_001()
 	assertEquals "Help calls" "1" "$HELP_CALLS"
 	assertEquals "Error calls" "0" "$ERROR_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
+	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 
 	setUp
 	pot-snapshot -h
@@ -37,7 +43,10 @@ test_pot_snapshot_001()
 	assertEquals "Help calls" "1" "$HELP_CALLS"
 	assertEquals "Error calls" "0" "$ERROR_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
+	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 
 	setUp
 	pot-snapshot -va
@@ -45,7 +54,10 @@ test_pot_snapshot_001()
 	assertEquals "Help calls" "1" "$HELP_CALLS"
 	assertEquals "Error calls" "1" "$ERROR_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
+	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 }
 
 test_pot_snapshot_002()
@@ -57,9 +69,11 @@ test_pot_snapshot_002()
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
 	assertEquals "_zfs_exist calls" "0" "$ZFSEXIST_CALLS"
 	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 	assertEquals "Info calls" "0" "$INFO_CALLS"
 }
 
@@ -72,9 +86,11 @@ test_pot_snapshot_003()
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
 	assertEquals "_zfs_exist calls" "0" "$ZFSEXIST_CALLS"
 	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 	assertEquals "Info calls" "0" "$INFO_CALLS"
 }
 
@@ -87,9 +103,11 @@ test_pot_snapshot_004()
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
 	assertEquals "_zfs_exist calls" "0" "$ZFSEXIST_CALLS"
 	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 	assertEquals "Info calls" "0" "$INFO_CALLS"
 }
 test_pot_snapshot_020()
@@ -102,7 +120,10 @@ test_pot_snapshot_020()
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
+	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 
 	setUp
 	pot-snapshot -p not-a-pot
@@ -112,7 +133,10 @@ test_pot_snapshot_020()
 	assertEquals "_is_pot calls" "1" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
+	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 
 	setUp
 	pot-snapshot -p test-pot-run
@@ -122,7 +146,10 @@ test_pot_snapshot_020()
 	assertEquals "_is_pot calls" "1" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "1" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
+	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 }
 
 test_pot_snapshot_021()
@@ -135,7 +162,10 @@ test_pot_snapshot_021()
 	assertEquals "_is_pot_running calls" "1" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "1" "$POTZFSSNAP_CALLS"
 	assertEquals "_pot_zfs_snap arg" "test-pot" "$POTZFSSNAP_CALL1_ARG1"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
+	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 }
 
 test_pot_snapshot_022()
@@ -147,8 +177,28 @@ test_pot_snapshot_022()
 	assertEquals "_is_pot calls" "1" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "1" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "1" "$POTZFSSNAPFULL_CALLS"
 	assertEquals "_pot_zfs_snap_full arg" "test-pot" "$POTZFSSNAPFULL_CALL1_ARG1"
+	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
+}
+
+test_pot_snapshot_023()
+{
+	pot-snapshot -p test-pot -r
+	assertEquals "Exit rc" "0" "$?"
+	assertEquals "Help calls" "0" "$HELP_CALLS"
+	assertEquals "Error calls" "0" "$ERROR_CALLS"
+	assertEquals "_is_pot calls" "1" "$ISPOT_CALLS"
+	assertEquals "_is_pot_running calls" "1" "$ISPOTRUN_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "1" "$RMVPOTSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap arg" "test-pot" "$RMVPOTSNAP_CALL1_ARG1"
+	assertEquals "_pot_zfs_snap calls" "1" "$POTZFSSNAP_CALLS"
+	assertEquals "_pot_zfs_snap arg" "test-pot" "$POTZFSSNAP_CALL1_ARG1"
+	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
+	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 }
 
 test_pot_snapshot_040()
@@ -160,7 +210,9 @@ test_pot_snapshot_040()
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 
 	setUp
 	pot-snapshot -f not-a-fscomp
@@ -170,9 +222,11 @@ test_pot_snapshot_040()
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
 	assertEquals "_zfs_exist calls" "1" "$ZFSEXIST_CALLS"
 	assertEquals "_fscomp_zfs_snap calls" "0" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 	
 }
 
@@ -185,8 +239,10 @@ test_pot_snapshot_041()
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
 	assertEquals "_zfs_exist calls" "1" "$ZFSEXIST_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 	assertEquals "_fscomp_zfs_snap calls" "1" "$FSCOMPZFSSNAP_CALLS"
 	assertEquals "_fscomp_zfs_snap arg" "test-fscomp" "$FSCOMPZFSSNAP_CALL1_ARG1"
 	assertEquals "_fscomp_zfs_snap arg" "" "$FSCOMPZFSSNAP_CALL1_ARG2"
@@ -202,8 +258,10 @@ test_pot_snapshot_042()
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
 	assertEquals "_zfs_exist calls" "1" "$ZFSEXIST_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 	assertEquals "_fscomp_zfs_snap calls" "1" "$FSCOMPZFSSNAP_CALLS"
 	assertEquals "_fscomp_zfs_snap arg" "test-fscomp" "$FSCOMPZFSSNAP_CALL1_ARG1"
 	assertEquals "_fscomp_zfs_snap arg" "" "$FSCOMPZFSSNAP_CALL1_ARG2"
@@ -219,11 +277,33 @@ test_pot_snapshot_043()
 	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
 	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
 	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
 	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
 	assertEquals "_zfs_exist calls" "1" "$ZFSEXIST_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "0" "$RMVFSCOMPSNAP_CALLS"
 	assertEquals "_fscomp_zfs_snap calls" "1" "$FSCOMPZFSSNAP_CALLS"
 	assertEquals "_fscomp_zfs_snap arg" "test-fscomp" "$FSCOMPZFSSNAP_CALL1_ARG1"
 	assertEquals "_fscomp_zfs_snap arg" "backup" "$FSCOMPZFSSNAP_CALL1_ARG2"
+	assertEquals "Info calls" "0" "$INFO_CALLS"
+}
+
+test_pot_snapshot_044()
+{
+	pot-snapshot -f test-fscomp -r
+	assertEquals "Exit rc" "0" "$?"
+	assertEquals "Help calls" "0" "$HELP_CALLS"
+	assertEquals "Error calls" "0" "$ERROR_CALLS"
+	assertEquals "_is_pot calls" "0" "$ISPOT_CALLS"
+	assertEquals "_is_pot_running calls" "0" "$ISPOTRUN_CALLS"
+	assertEquals "_pot_zfs_snap calls" "0" "$POTZFSSNAP_CALLS"
+	assertEquals "_remove_oldest_pot_snap calls" "0" "$RMVPOTSNAP_CALLS"
+	assertEquals "_pot_zfs_snap_full calls" "0" "$POTZFSSNAPFULL_CALLS"
+	assertEquals "_zfs_exist calls" "1" "$ZFSEXIST_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap calls" "1" "$RMVFSCOMPSNAP_CALLS"
+	assertEquals "_remove_oldest_fscomp_snap arg" "test-fscomp" "$RMVFSCOMPSNAP_CALL1_ARG1"
+	assertEquals "_fscomp_zfs_snap calls" "1" "$FSCOMPZFSSNAP_CALLS"
+	assertEquals "_fscomp_zfs_snap arg" "test-fscomp" "$FSCOMPZFSSNAP_CALL1_ARG1"
+	assertEquals "_fscomp_zfs_snap arg" "" "$FSCOMPZFSSNAP_CALL1_ARG2"
 	assertEquals "Info calls" "0" "$INFO_CALLS"
 }
 
@@ -233,6 +313,8 @@ setUp()
 	HELP_CALLS=0
 	FSCOMPZFSSNAP_CALL1_ARG1=""
 	FSCOMPZFSSNAP_CALL1_ARG2=""
+	RMVPOTSNAP_CALL1_ARG1=""
+	RMVFSCOMPSNAP_CALL1_ARG1=""
 }
 
 . shunit/shunit2